% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsmixGau.R
\name{BICFunction}
\alias{BICFunction}
\title{Model Selection Via Bayesian Information Criterion}
\usage{
BICFunction(
  logLikelihood,
  nParameters,
  nObservations,
  clusterRunOutput = NA,
  gmin,
  gmax
)
}
\arguments{
\item{logLikelihood}{A vector with value of final log-likelihoods for
each cluster size.}

\item{nParameters}{A vector with number of parameters for each
cluster size.}

\item{nObservations}{A positive integer specifying the number of observations
in the dataset analyzed.}

\item{clusterRunOutput}{Output from mplnVariational, mplnMCMCParallel, or
mplnMCMCNonParallel, if available. Default value is NA. If provided,
the vector of cluster labels obtained by mclust::map() for best model
will be provided in the output.}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >gmin, specifying the maximum number of
components to be considered in the clustering run.}
}
\value{
Returns an S3 object of class MPLN with results.
\itemize{
\item allBICvalues - A vector of BIC values for each cluster size.
\item BICmodelselected - An integer specifying model selected by BIC
\item BICmodelSelectedLabels - A vector of integers specifying cluster labels
for the model selected. Only provided if user input clusterRunOutput.
\item BICMessage - A character vector indicating if spurious clusters are
detected. Otherwise, NA.
}
}
\description{
Performs model selection using Bayesian Information Criterion (BIC) by
Schwarz (1978). Formula: - 2 * logLikelihood + (nParameters * log(nObservations)).
}
\examples{

# Generating simulated data
G = 2 # number of true clusters/components
dimension = 6
nObservations = 1000
piGTrue <- c(0.8, 0.2)

set.seed(1234)
mean1 <- rep(1, dimension)
mean2 <- rep(4, dimension)
sigma1 <- diag(dimension) * 0.2
sigma2 <- diag(dimension) * 0.5

library(mvtnorm)
component1 <- mvtnorm::rmvnorm(nObservations * piGTrue[1], mean = mean1, sigma = sigma1)
component2 <- mvtnorm::rmvnorm(nObservations * piGTrue[2], mean = mean2, sigma = sigma2)
dataset <- rbind(component1, component2)

# Cluster data
clustOutput <- mixGaussianEM(dataset = dataset,
                             membership = "none",
                             gmin = 1,
                             gmax = 2,
                             initMethod = "kmeans",
                             nInitIterations = 1)

# Model selection
BICmodel <- BICFunction(logLikelihood = clustOutput$logLikelihood,
                         nParameters = clustOutput$numbParameters,
                         nObservations = nrow(clustOutput$dataset),
                         clusterRunOutput = clustOutput$allResults,
                         gmin = clustOutput$gmin,
                         gmax = clustOutput$gmax)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.
}
\author{
{Anjali Silva, \email{anjali.silva@uhnresearch.ca}}
}
