% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsmixGau.R
\name{ICLFunction}
\alias{ICLFunction}
\title{Model Selection Via Integrated Completed Likelihood}
\usage{
ICLFunction(
  logLikelihood,
  nParameters,
  nObservations,
  clusterRunOutput = NA,
  probaPost = NA,
  gmax,
  gmin
)
}
\arguments{
\item{logLikelihood}{A vector with value of final log-likelihoods for
each cluster size.}

\item{nParameters}{A vector with number of parameters for each
cluster size.}

\item{nObservations}{A positive integer specifying the number of observations
in the dataset analyzed.}

\item{clusterRunOutput}{Output from mixGaussianEM function. Either clusterRunOutput
or probaPost must be provided.}

\item{probaPost}{A list that is length (gmax - gmin + 1) containing posterior
probability at each g, for g = gmin:gmax. This argument is useful if
clustering output have been generated non-serially, e.g., g = 1:5 and
g = 6:10 rather than g = 1:10. Either clusterRunOutput or probaPost
must be provided.}

\item{gmax}{A positive integer, > gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}
}
\value{
Returns an S3 object of class MPLN with results.
\itemize{
\item allICLvalues - A vector of ICL values for each cluster size.
\item ICLmodelselected - An integer specifying model selected by ICL.
\item ICLmodelSelectedLabels - A vector of integers specifying cluster labels
for the model selected. Only provided if user input clusterRunOutput.
\item ICLMessage - A character vector indicating if spurious clusters are
detected. Otherwise, NA.
}
}
\description{
Performs model selection using integrated completed likelihood (ICL) by
Biernacki et al., (2000).
}
\examples{

# Generating simulated data
G = 2 # number of true clusters/components
dimension = 6
nObservations = 1000
piGTrue <- c(0.8, 0.2)

set.seed(1234)
mean1 <- rep(1, dimension)
mean2 <- rep(4, dimension)
sigma1 <- diag(dimension) * 0.2
sigma2 <- diag(dimension) * 0.5

# library(mvtnorm)
component1 <- mvtnorm::rmvnorm(nObservations * piGTrue[1], mean = mean1, sigma = sigma1)
component2 <- mvtnorm::rmvnorm(nObservations * piGTrue[2], mean = mean2, sigma = sigma2)
dataset <- rbind(component1, component2)

# Cluster data
clustOutput <- mixGaussian::mixGaussianEM(dataset = dataset,
                                          membership = "none",
                                          gmin = 1,
                                          gmax = 2,
                                          initMethod = "kmeans",
                                          nInitIterations = 1)

# Model selection
ICLmodel <- mixGaussian::ICLFunction(logLikelihood = clustOutput$logLikelihood,
                                     nParameters = clustOutput$numbParameters,
                                     nObservations = nrow(clustOutput$dataset),
                                     clusterRunOutput = clustOutput$allResults,
                                     gmin = clustOutput$gmin,
                                     gmax = clustOutput$gmax)

}
\references{
Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions
on Pattern Analysis and Machine Intelligence} 22.
}
\author{
{Anjali Silva, \email{anjali.silva@uhnresearch.ca}}
}
