% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsClassification.R
\name{mixGaussianClassEM}
\alias{mixGaussianClassEM}
\title{Classification Using Mixtures of Multivariate Gaussian Via EM}
\usage{
mixGaussianClassEM(
  dataset,
  membership = "none",
  gmin,
  gmax,
  initMethod = "random",
  nInitIterations = 2
)
}
\arguments{
\item{dataset}{A dataset of class matrix such that rows correspond to
observations and columns correspond to variables. The dataset have
dimensions n x d, where n is the total number of observations and d
is the dimensionality. If rowSums are zero, these rows will be removed
prior to cluster analysis.}

\item{membership}{A numeric vector of length nrow(dataset) containing the
cluster membership of each observation. If not available, leave as "none"
and model-based clustering will be performed. If a vector provided, then
model-based classification is performed. In this latter case, the ith entry
of membership is either zero, indicating that the component membership of
observation i is unknown, or it corresponds to the component membership of
observation i. See Examples below.}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >= gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{initMethod}{An algorithm for initialization. Current options are
"kmeans", "random", "medoids", "clara", or "fanny". Default is "kmeans".}

\item{nInitIterations}{A positive integer or zero, specifying the number
of initialization runs to be performed. This many runs, each with 10
iterations, will be performed via mixGaussianClust and values from
the run with highest log-likelihood will be used as initialization
values. Default is 2.}
}
\value{
Returns an S3 object of class mixGaussianClass with results.
\itemize{
\item dataset - The input dataset on which clustering is performed.
\item dimensionality - Dimensionality of the input dataset.
\item gmin - Minimum number of components/clusters considered in the clustering
run.
\item gmax - Maximum number of components/clusters considered in the clustering
run.
\item initalizationMethod - Method used for initialization.
\item allResults - A list with all results.
\item logLikelihood - A vector with value of final log-likelihoods for
each component/cluster size.
\item numbParameters - A vector with number of parameters for each
component/cluster size.
\item trueLabels - The vector of true labels, if provided by user.
\item ICLresults - A list with all ICL model selection results.
\item BICresults - A list with all BIC model selection results.
\item AICresults - A list with all AIC model selection results.
\item AIC3results - A list with all AIC3 model selection results.
\item totalTime - Total time used for clustering and model selection.
}
}
\description{
Performs classification using mixtures of multivariate Gaussian
distribution with expectation-maximization (EM) for parameter
estimation. Model selection is performed using AIC, AIC3,
BIC and ICL.
}
\examples{
# Example 1: Generating simulated data
set.seed(1234)
G <- 2 # number of true clusters/components
dimension <- 6
nObservations <- 100
membershipTrue <- sample(c(1, 2), 100, replace = TRUE)
piGTrue <- table(membershipTrue)/nObservations

# Simulate parameter values
mean1 <- rep(1, dimension)
mean2 <- rep(4, dimension)
sigma1 <- diag(dimension) * 2
sigma2 <- diag(dimension) * 2

# library(mvtnorm)
component1 <- mvtnorm::rmvnorm(nObservations * piGTrue[1], mean = mean1, sigma = sigma1)
dim(component1)
component2 <- mvtnorm::rmvnorm(nObservations * piGTrue[2], mean = mean2, sigma = sigma2)
dim(component2)
dataset <- rbind(component1, component2)
dim(dataset) # 100   6

# Visualize data
# pairs(dataset, col = c(rep(2, nObservations * piGTrue[1]), rep(3, nObservations * piGTrue[2])))

# Classification membership vector
membershipClass <- membershipTrue
membershipClass[x = sample(c(1:100),
                size = 10,
                replace = FALSE)] <- 0
table(membershipClass)
#  0  1  2
# 10 44 46

# Classify data, where cluster membership of observations
# with membership indicated by 0 are unknown
clustOutput <- mixGaussian::mixGaussianEM(dataset = dataset,
                                          membership = membershipClass,
                                          gmin = 1,
                                          gmax = 5,
                                          initMethod = "kmeans",
                                          nInitIterations = 1)


# Example 2: Access raw data made available with this package
# Not run
# inputCountsPath <- system.file("extdata", "mixGaussianDataset.csv",
#                               package = "mixGaussian")
# Read data
# exampleData <- read.csv(file = inputCountsPath, header = TRUE)
# dim(exampleData) # 1000  6
# To see documentation for this dataset
?exampleData

# Cluster data
# clustOutput <- mixGaussian::mixGaussianEM(dataset = exampleData,
#                                          membership = "none",
#                                          gmin = 1,
#                                          gmax = 5,
#                                          initMethod = "kmeans",
#                                          nInitIterations = 1)

}
\references{
Aitken, A. C. (1926). A series formula for the roots of algebraic and transcendental equations.
\emph{Proceedings of the Royal Society of Edinburgh}, 45, 14–22.

B¨ohning, D., E. Dietz, R. Schaub, P. Schlattmann, and B. Lindsay (1994). The distribution
of the likelihood ratio for mixtures of densities from the one-parameter exponential family.
\emph{Annals of the Institute of Statistical Mathematics}, 46, 373–388.

Dempster, A. P., N. M. Laird, and D. B. Rubin (1977). Maximum likelihood from incomplete
data via the EM algorithm. \emph{Journal of the Royal Statistical Society: Series B} 39, 1–38.
}
\author{
{Anjali Silva, \email{anjali@alumni.uoguelph.ca}}
}
