% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsmixGau.R
\name{mixGaussianEM}
\alias{mixGaussianEM}
\title{Clustering Using Mixtures of Multivariate Gaussian Via EM}
\usage{
mixGaussianEM(
  dataset,
  membership = "none",
  gmin,
  gmax,
  initMethod = "random",
  nInitIterations = 2
)
}
\arguments{
\item{dataset}{A dataset of class matrix such that rows correspond to
observations and columns correspond to variables. The dataset have
dimensions n x d, where n is the total number of observations and d
is the dimensionality. If rowSums are zero, these rows will be removed
prior to cluster analysis.}

\item{membership}{A numeric vector of length nrow(dataset) containing the
cluster membership of each observation. If not available, leave as "none".}

\item{gmin}{A positive integer specifying the minimum number of components
to be considered in the clustering run.}

\item{gmax}{A positive integer, >= gmin, specifying the maximum number of
components to be considered in the clustering run.}

\item{initMethod}{An algorithm for initialization. Current options are
"kmeans", "random", "medoids", "clara", or "fanny". Default is "kmeans".}

\item{nInitIterations}{A positive integer or zero, specifying the number
of initialization runs to be performed. This many runs, each with 10
iterations, will be performed via mixGaussianClust and values from
the run with highest log-likelihood will be used as initialization
values. Default is 2.}
}
\value{
Returns an S3 object of class mplnVariational with results.
\itemize{
\item dataset - The input dataset on which clustering is performed.
\item dimensionality - Dimensionality of the input dataset.
\item gmin - Minimum number of components/clusters considered in the clustering
run.
\item gmax - Maximum number of components/clusters considered in the clustering
run.
\item initalizationMethod - Method used for initialization.
\item allResults - A list with all results.
\item logLikelihood - A vector with value of final log-likelihoods for
each component/cluster size.
\item numbParameters - A vector with number of parameters for each
component/cluster size.
\item trueLabels - The vector of true labels, if provided by user.
\item ICLresults - A list with all ICL model selection results.
\item BICresults - A list with all BIC model selection results.
\item AICresults - A list with all AIC model selection results.
\item AIC3results - A list with all AIC3 model selection results.
\item totalTime - Total time used for clustering and model selection.
}
}
\description{
Performs clustering using mixtures of multivariate Gaussian
distribution with expectation-maximization (EM) for parameter
estimation. Model selection is performed using AIC, AIC3,
BIC and ICL.
}
\examples{
# Generating simulated data
G = 2 # number of true clusters/components
dimension = 6
nObservations = 1000
piGTrue <- c(0.8, 0.2)

set.seed(1234)
mean1 <- rep(1, dimension)
mean2 <- rep(4, dimension)
sigma1 <- diag(dimension) * 2
sigma2 <- diag(dimension) * 2

library(mvtnorm)
component1 <- mvtnorm::rmvnorm(nObservations * piGTrue[1], mean = mean1, sigma = sigma1)
dim(component1) # 52   6
component2 <- mvtnorm::rmvnorm(nObservations * piGTrue[2], mean = mean2, sigma = sigma2)
dim(component2) # 48   6
dataset <- rbind(component1, component2)
dim(dataset) # 100   6
pairs(dataset, col = c(rep(2, nObservations * piGTrue[1]), rep(3, nObservations * piGTrue[2])))

clustOutput <- mixGaussianEM(dataset = dataset,
                             membership = "none",
                             gmin = 1,
                             gmax = 2,
                             initMethod = "random",
                             nInitIterations = 1)


}
\references{

}
\author{
{Anjali Silva, \email{anjali.silva@uhnresearch.ca}}
}
