% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsmixGau.R
\name{runMixGaussian}
\alias{runMixGaussian}
\title{Launch Shiny App For Package mixGaussian}
\usage{
runMixGaussian()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app permit to perform clustering using mixtures
of multivariate Gaussian distributions via EM. Model selection
is performed using AIC, AIC3, BIC and ICL.
}
\examples{
\dontrun{
runMixGaussian()
}

}
\references{
Aitken, A. C. (1926). A series formula for the roots of algebraic and transcendental equations.
\emph{Proceedings of the Royal Society of Edinburgh}, 45, 14–22.

B¨ohning, D., E. Dietz, R. Schaub, P. Schlattmann, and B. Lindsay (1994). The distribution
of the likelihood ratio for mixtures of densities from the one-parameter exponential family.
\emph{Annals of the Institute of Statistical Mathematics}, 46, 373–388.

Dempster, A. P., N. M. Laird, and D. B. Rubin (1977). Maximum likelihood from incomplete
data via the EM algorithm. \emph{Journal of the Royal Statistical Society: Series B} 39, 1–38.
}
\author{
Anjali Silva, \email{anjali.silva@uhnresearch.ca}
}
